requirements.txt
PySide2.QtCore,
PySide2.QtWidgets,
PySide2.QtGui